% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_annotation}
\alias{add_annotation}
\title{Wrapper around ggplot2::annotation_custom for simplified annotation to ggplot2 plots}
\usage{
add_annotation(
  gg = NULL,
  lbl = NULL,
  base_family = "sans",
  base_size = 11,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf
)
}
\arguments{
\item{gg}{Object of class \code{ggplot}.}

\item{lbl}{\code{String, dataframe, data.table, tibble} used to annotate the \code{ggplot}.}

\item{base_family}{\code{character}. Base font family}

\item{base_size}{\code{numeric}. Base font size in pt}

\item{xmin, }{xmax x coordinates giving horizontal location of raster in which to fit annotation.}

\item{ymin, }{ymax y coordinates giving vertical location of raster in which to fit annotation.}
}
\value{
Object of class \code{ggplot} \code{ggsurvfit} with added annotation with an object of class \code{gtable}.
}
\description{
This function accepts a string, dataframe, data.table, tibbles and customized objects of class \code{gtable} and
  places them on the specified location on the \code{ggplot}. The layout is fixed: bold columnheaders and plain body. Only the
  font size and type can be chosen.
}
\examples{

## Estimate survival
surv_object <- vr_KM_est(data = adtte, strata = "TRTP")

## p value for "Equality across strata"
lbl <- get_pvalue(survfit_object, statlist = c("test", "p"), type = "All")

## add results to survival plot
vr_plot(survfit_object) \%>\%
  add_annotation(lbl = lbl, base_family = "sans", base_size = 9, xmin = 110, xmax = 180, ymin = 0.80)

}
\seealso{
\code{\link[gridExtra]{tableGrob}} \code{\link[ggplot2]{annotation_custom}}
}
\author{
Steven Haesendonckx {shaesen2@its.jnj.com}
}
