% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\alias{add_CI.ggsurvfit}
\title{Add pointwise Confidence Interval to a an object created by visR through an S3 method}
\usage{
add_CI(gg, ...)

\method{add_CI}{ggsurvfit}(gg, alpha = 0.1, style = "ribbon", linetype, ...)
}
\arguments{
\item{gg}{A ggplot created with visR}

\item{...}{other arguments passed on to the method to modify \code{\link[ggplot2]{geom_ribbon}}}

\item{alpha}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is 0.1.}

\item{style}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is "ribbon". An alternative option is "step" that uses a line to display interval bounds.}

\item{linetype}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}.}
}
\value{
Pointwise confidence interval overlayed on a visR ggplot
}
\description{
S3 method for adding a pointwise confidence interval to a object created with visR.
No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(visR)

## plot without confidence intervals
p <- visR::plot(survfit_object) 
p

# plot with default settings
p \%>\% add_CI()

# change transparency of ribbon  
p \%>\% add_CI(alpha = 0.9, style = "ribbon")

# plot interval as a step
p \%>\% add_CI(alpha = 0.1, style = "step")

change linetype 
p \%>\% add_CI(style = "step", linetype = 1)

}
\author{
Steven Haesendonckx
}
