% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_highlight.R
\name{add_highlight}
\alias{add_highlight}
\alias{add_highlight.ggsurvfit}
\title{Highlight a specific strata}
\usage{
add_highlight(gg, ...)

\method{add_highlight}{ggsurvfit}(gg, strata, new_alpha = 0.5, ...)
}
\arguments{
\item{gg}{A ggplot created with visR}

\item{...}{other arguments passed on to the method to modify \code{\link[ggplot2]{geom_ribbon}}}

\item{strata}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is 0.1.}

\item{alpha}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is "ribbon".}
}
\value{
Pointwise confidence interval overlayed on a visR ggplot
}
\description{
S3 method for adding a pointwise confidence interval to a object created with visR.
No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
vr_plot(survfit_object) \%>\%
  add_CI(alpha = 0.1, style = "step", linetype = 3)

}
\author{
Tim Treis
}
