% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{add_metadata_gt}
\alias{add_metadata_gt}
\title{Internal function to add metadata to a gt object}
\usage{
add_metadata_gt(gt, title, datasource, footnote)
}
\arguments{
\item{gt}{input gt object}

\item{title}{Specify the title as a text string to be displayed in the rendered table.
Default is no title.}

\item{datasource}{String specifying the data source underlying the data set.
Default is no title.}

\item{footnote}{String specifying additional information to be displayed as a footnote
alongside the data source and specifications of statistical tests.}
}
\value{
gt object
}
\description{
Internal function to add metadata to a gt object
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
