% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantiles.R
\name{add_quantiles}
\alias{add_quantiles}
\alias{add_quantiles.ggsurvfit}
\title{Add lines indicating the survival quantiles to a survival plot}
\usage{
add_quantiles(gg, ...)

\method{add_quantiles}{ggsurvfit}(
  gg,
  quantiles = 0.5,
  linetype = "mixed",
  linecolour = "strata",
  alpha = 1,
  ...
)
}
\arguments{
\item{gg}{A ggplot created with visR}

\item{...}{other arguments passed on to the method to modify \code{\link[ggplot2]{geom_line}}}

\item{quantiles}{y-axis values for which to draw the lines (default: 0.5)}

\item{linetype}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_line}} (default: dashed)}

\item{linecolour}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_line}} (default: black)}

\item{alpha}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_line}} (default: 0.8)}
}
\value{
Lines indicating the quantiles overlayed on a visR ggplot
}
\description{
Method to add lines indicating the specified quantiles of the survival curve.
}
\examples{

library(visR)

adtte \%>\%
  estimate_KM("SEX") \%>\%
  visr() \%>\%
  add_quantiles() 

adtte \%>\%
  estimate_KM("SEX") \%>\%
  visr() \%>\%
  add_quantiles(c(0.25, 0.50, 0.75)) 

adtte \%>\%
  estimate_KM("SEX") \%>\%
  visr() \%>\%
  add_quantiles(c(0.25, 0.50, 0.75)) 

}
