% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable.R
\name{add_risktable}
\alias{add_risktable}
\alias{add_risktable.ggsurvfit}
\title{Add risk tables to visR plots through an S3 method}
\usage{
add_risktable(
  gg,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = "At risk",
  group = "strata",
  collapse = FALSE
)

\method{add_risktable}{ggsurvfit}(
  gg,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = "At risk",
  group = "strata",
  collapse = FALSE
)
}
\arguments{
\item{gg}{visR plot of class \code{ggsurvfit}}

\item{min_at_risk}{\code{numeric} The cutoff for number of subjects to display. Default is 0.}

\item{breaks}{Single numeric or numeric vector indicating breaks.}

\item{statlist}{Character vector indicating which summary data to present. Current choices are "n.risk" "n.event" "n.censor".}

\item{label}{Character vector with labels for the statlist. Default matches "n.risk" with "At risk", "n.event" with "Events" and "n.censor"
with "Censored".}

\item{group}{String indicating the grouping variable for the risk tables. Current options are:
\itemize{
\item{"strata": groups the risk tables per stratum. The \code{label} specifies the lables used within each risk table. This is the default}
\item{"statlist": groups the risk tables per statlist. The \code{label} specifies the title for each risk tabel. The strata levels
are used for labeling within each risk table.}
} "strata" to group the risk tables
per strata, and "statlist" to group the risk tables}

\item{collapse}{Boolean, indicates whether to present the data overall, rather than per strata.
Default is FALSE.}
}
\value{
Object of class \code{ggplot} with added risk table.
}
\description{
S3 method for adding risk tables to visR plots.
No default method is available at the moment.
}
\examples{
\donttest{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)
library(gtable)
library(visR)

## Display 2 risk tables, 1 per statlist
adtte \%>\%
  estimate_KM(strata = "TRTP") \%>\%
  visR::plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,group = "statlist"
                )
                
## Display 2 risk tables, 1 per stratum
adtte \%>\%
  estimate_KM(strata = "TRTP") \%>\%
  visR::plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,group = "strata"
                )

## Display overall risk table
adtte \%>\%
  estimate_KM(strata = "TRTP") \%>\%
  visR::plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,collapse = TRUE
                )

}

}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
