% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca_cohort}
\alias{brca_cohort}
\title{Cancer survival data}
\format{
An object of class \code{data.frame} with 1098 rows and 10 columns.
}
\usage{
brca_cohort
}
\description{
Creation script in data-raw
}
\keyword{datasets}
