% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{check_rendering_input}
\alias{check_rendering_input}
\title{Internal function to check if the input works}
\usage{
check_rendering_input(output_format = NULL, engine = NULL)
}
\arguments{
\item{output_format}{format for output i.e. html}

\item{engine}{engine to render output.}
}
\value{
Warning message
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
