% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_helpers.R
\name{check_traceback_stack_for_ggplot_aesthetics_warning}
\alias{check_traceback_stack_for_ggplot_aesthetics_warning}
\title{Collections of functions only used in testthat tests}
\usage{
check_traceback_stack_for_ggplot_aesthetics_warning()
}
\description{
Collections of functions only used in testthat tests
}
\section{Last updated by}{

Tim Treis
}

\section{Last update date}{

18-MAR-2021
A helper function for catching errors in the ggplot2 traceback stack
}

\keyword{internal}
