% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Data from Main Experiment}
\format{A data frame with many rows and columns
\describe{
\item{participant}{ number}
\item{trial}{Trial number, from 1..NNN}
...
}}
\usage{
d
}
\description{
Pretty great and collected data
}
\keyword{datasets}
