% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{.convert_alpha}
\alias{.convert_alpha}
\title{Converts an alpha value between its numeric and its hex-encoded form.}
\usage{
.convert_alpha(numeric_alpha = NULL, hex_alpha = NULL)
}
\arguments{
\item{numeric_alpha}{A numerical value between 0 and 1.}

\item{hex_alpha}{A two-letter character string.}
}
\value{
If \code{numeric_alpha} was specified, its two-letter representation is returned. If \code{hex_alpha} was specified, its numeric representation is returned.
}
\description{
The function accepts a numeric (or NULL/NA) or a two-character hex encoded alpha representation and returns a the respective other representation.
}
\keyword{internal}
