% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{.get_coordinates}
\alias{.get_coordinates}
\title{Create coordinates for each row in the attrition table}
\usage{
.get_coordinates(data, box_width, box_height, field_height)
}
\arguments{
\item{data}{A dataframe containing the attrition data}

\item{box_width}{\code{integer} The width of the boxes in the flow charts (in canvas coordinates)}

\item{box_height}{\code{integer} The height of the boxes in the flow charts (in canvas coordinates)}

\item{field_height}{\code{float} The width of the boxes in the flow charts (in canvas coordinates)}
}
\value{
The input dataframe extended by columns containing x and y coordinates for included and excluded counts
}
\description{
This function creates lables with a maximal character length per line by combining content of two dataframe columns
}
\keyword{internal}
