% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{.get_labels}
\alias{.get_labels}
\title{Create labels for flowchart}
\usage{
.get_labels(
  data,
  description_column_name,
  value_column_name,
  complement_column_name = "",
  wrap_width = 50
)
}
\arguments{
\item{data}{A dataframe}

\item{description_column_name}{\code{character} The column name containing description part of the new label}

\item{value_column_name}{\code{character} The column name containing the sample size part of the new label}

\item{complement_column_name}{\code{character} The column name containing a complement description part (will result in a second label)}

\item{wrap_width}{\code{integer} for the maximal character count per line}
}
\value{
The input dataframe extended by two columns containing the label and complement label
}
\description{
This function creates lables with a maximal character length per line by combining content of two dataframe columns
}
\keyword{internal}
