% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{.get_labelsizes}
\alias{.get_labelsizes}
\title{Calculate the size labels on the}
\usage{
.get_labelsizes(data, label, complement_label)
}
\arguments{
\item{data}{Dataframe with label and complement label strings}

\item{label}{The column containing attrition labels}

\item{complement_label}{The column containing complement description labels}
}
\value{
The input dataframe extended by several columns containing the label and complement label height and width
}
\description{
Calculate the text width and maximal text width of both the label and complement labels
}
\keyword{internal}
