% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{.get_strata}
\alias{.get_strata}
\title{Get strata level combinations}
\usage{
.get_strata(strata)
}
\arguments{
\item{strata}{The strata from a survival object}
}
\value{
The strata level combinations from a survival objects without the specified strata.
}
\description{
Extracts the strata level combinations from a survival objects without the specified strata.
}
\keyword{internal}
