% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{.pvalformat}
\alias{.pvalformat}
\title{Internal function to format pvalues.}
\usage{
.pvalformat(x)
}
\arguments{
\item{x}{Input p-value. Numeric value.}
}
\description{
Internal function to format pvalues.
}
