% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_CUMINC.R
\name{estimate_CUMINC}
\alias{estimate_CUMINC}
\title{Competing Events Cumulative Incidence}
\usage{
estimate_CUMINC(data, strata = NULL, CNSR, AVAL, conf.int = 0.95, ...)
}
\arguments{
\item{data}{The name of the dataset for Time-to-Event analysis based on the Analysis Data Model (ADaM) principles. The dataset is expected to have
one record per subject per analysis parameter. Rows in which the analysis variable (AVAL) or the sensor variable (CNSR) contain NA, are removed during analysis.}

\item{strata}{Character vector, representing the strata for Time-to-Event analysis. When NULL, an overall analysis is performed.
Default is NULL.}

\item{CNSR}{Column name indicating the outcome and censoring statuses.
Column must be a factor and the first level indicates censoring, the
next level is the outcome of interest, and the remaining levels are the
competing events.}

\item{AVAL}{Analysis value for Time-to-Event analysis. Default is "AVAL", as per CDISC ADaM guiding principles.}

\item{conf.int}{Confidence internal level. Default is 0.95.}

\item{...}{additional arguments passed on to the ellipsis of the call \code{survival::survfit.formula(data = data, formula = Surv(AVAL, 1-CNSR) ~ strata), ...)} .
Use \code{?survival::survfit.formula} and \code{?survival::survfitCI} for more information.}
}
\description{
Competing Events Cumulative Incidence
}
\examples{
estimate_CUMINC(
  tidycmprsk::trial,
  strata = "trt",
  CNSR = "death_cr",
  AVAL = "ttdeath"
) \%>\%
  visr() \%>\%
  add_CI() \%>\%
  add_risktable(statlist = c("n.risk", "cumulative.event"))
}
