% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{get_alpha_from_hex_colour}
\alias{get_alpha_from_hex_colour}
\title{Extract the numerical alpha representation of #RRGGBBAA colour}
\usage{
get_alpha_from_hex_colour(hex_colour = NULL)
}
\arguments{
\item{hex_colour}{A string in the format '#RRGGBBAA'}
}
\value{
The numeric representation of the colors' alpha value, rounded to 2 digits.
}
\description{
The colours for the strata are usually encoded by their hex-representation. This yields a string with 3 blocks of hex-encoded values between 0-255 for the colours red, green and blue in the form of '#RRGGBB'. Additionally, a fourth optional block can be present encoding the alpha transparency of the colour. This extends the string to '#RRGGBBAA'. This function extracts the 'AA' part of a colour and returns the numerical representation if it.
}
\keyword{internal}
