% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_COX_HR.R
\name{get_COX_HR}
\alias{get_COX_HR}
\alias{get_COX_HR.survfit}
\title{Summarize Hazard Ratio from a survival object using S3 method}
\usage{
get_COX_HR(x, ...)

\method{get_COX_HR}{survfit}(survfit_object = survfit_object, update_formula = NULL, ...)
}
\arguments{
\item{x}{S3 object}

\item{...}{other arguments passed on to the method survival::coxph}

\item{survfit_object}{An object of class \code{survfit}}

\item{update_formula}{Template which specifies how to update the formula of the survfit_object \code{\link[stats]{update.formula}}}
}
\value{
A tibble with Hazard Ratios
}
\description{
S3 method for extracting information regarding Hazard Ratios.
    No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

## treatment effect
survfit_object_trt <- vr_KM_est(data = adtte, strata = c("TRTP"))
get_COX_HR(survfit_object_trt)

## treatment and gender effect
survfit_object_trt_sex <- vr_KM_est(data = adtte, strata = c("TRTP", "SEX"))
get_COX_HR(survfit_object_trt_sex)

## update formula of KM estimates by treatment to include "SEX" for HR estimation
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + SEX")
 
## update formula of KM estimates by treatment to include "AGE" for HR estimation with ties considered via the efron method
get_COX_HR(survfit_object_trt, update_formula = ". ~ . + strata(AGE)", ties = "efron")

}
\seealso{
\code{\link[survival]{coxph}} \code{\link[stats]{update.formula}}
}
\author{
Steven Haesendonckx
}
