% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{get_gt.data.frame}
\alias{get_gt.data.frame}
\title{Internal function Get gt object for a data.frame}
\usage{
\method{get_gt}{data.frame}(data, numcols)
}
\arguments{
\item{data}{input data set}

\item{numcols}{number of columns}
}
\value{
gt object
}
\description{
Internal function Get gt object for tableone
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
