% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tableone.R
\name{get_tableone}
\alias{get_tableone}
\alias{get_tableone.default}
\title{#' Create Summary Table (also known as 'tableone')}
\usage{
get_tableone(
  data,
  strata = NULL,
  overall = TRUE,
  summary_function = summarize_tab1
)

\method{get_tableone}{default}(
  data,
  strata = NULL,
  overall = TRUE,
  summary_function = summarize_tab1
)
}
\arguments{
\item{data}{The dataset to summarize as dataframe or tibble}

\item{strata}{Stratifying/Grouping variable name(s) as character vector. If NULL, only overall results are returned}

\item{overall}{If TRUE, the summary statistics for the overall dataset are also calculated}

\item{summary_function}{A function defining summary statistics for numeric and categorical values}
}
\description{
S3 method for creating summary tables (tableone).

By default the following summary stats are calculated:
\itemize{
\item Numeric variables: mean, min, 25th-percentile, median, 75th-percentile, maximum, standard deviation
\item Factor variables: proportion of each factor level in the overall dataset
\item Default: number of unique values and number of missing values
}
}
\details{
It is possible to provide your own summary function. Please have a loot at summary for inspiration.
}
\note{
All columns in the table will be summarized. If only some columns shall be used, please select only those
variables prior to creating the summary table by using dplyr::select()
}
\examples{
library(survival)
library(dplyr)
ovarian \%>\% 
select(-fustat) \%>\% 
  mutate(age_group = factor(case_when(age <= 50 ~ "<= 50 years",
                                      age <= 60 ~ "<= 60 years",
                                      age <= 70 ~ "<= 70 years",
                                      TRUE ~ "> 70 years")),
         rx = factor(rx),
         ecog.ps = factor(ecog.ps)) \%>\% 
  select(age, age_group, everything()) \%>\% 
  get_tableone()
    
}
