% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_sequence.R
\name{get_tx_sequence}
\alias{get_tx_sequence}
\title{Generate treatment sequence summaries from one-row-per-patient-line data}
\usage{
get_attrition(data, id, label, line, n_groups)
}
\arguments{
\item{data}{\code{data.frame} containing one-row-per-patient-line.
The data.frame must contain columns for 1) a patient identifier
(character or integer), 2) a treatment label (character or integer), and
3) a line number (integer or factor for line labels). Optionally, a fourth
column specifying the IDs of patients can be included to display reasons for
censorship.

See description for more information.}

\item{id}{\code{character/integer} the column name of the patient id.}

\item{label}{\code{character} specifying the column name of the
treatment or treatment group.}

\item{line}{\code{integer/factor} specifying the column name of the line
number (or line label) columns.}

\item{mortality_ids}{\code{character/integer} (optional) vector specifying
which patients should be displayed as dead instead of censored.}

\item{n_groups}{\code{integer} the maximum number of unique
treatment labels that will be considered (others will be grouped as "other").
Default is 6.}
}
\value{
Treatment sequencing data in lodes form.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This is an experimental function that may be developed over time.

This function manipulates treatment sequence information given
as one-row-per-line into "lodes" form, one row per combination. The outputs
of this function can be used as inputs into visr() to create a
sankey plot.
}
\examples{

lots <- tibble(
  patient = c('abc','abc','xyz','xyz','xyz'),
  line_number = c(1,2,1,2,3),
  treatment = c('Drug 1','Drug 3', 'Drug 2', 'Drug 4','Drug 3')
)

visR::get_attrition(lots,
  id = 'patient',
  label = 'treatment',
  line = 'line_number'
  )


}
