% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_helpers.R
\name{map_numbers_to_new_range}
\alias{map_numbers_to_new_range}
\title{A helper function for map numbers to an arbitrary range}
\usage{
map_numbers_to_new_range(numbers, lower, upper)
}
\description{
A helper function for map numbers to an arbitrary range
}
\keyword{internal}
