% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{numeric_alpha_to_hex}
\alias{numeric_alpha_to_hex}
\title{Converts a the numeric representation of an alpha value into the hex-encoded version.}
\usage{
numeric_alpha_to_hex(numeric_alpha = NULL)
}
\arguments{
\item{numeric_alpha}{A numerical value between 0 and 1.}
}
\value{
A two-character sequence that represents the hex-value of \code{numeric_alpha}.
}
\description{
The function accepts a numeric value (or NULL/NA) and returns a two-character hex representation of it.
}
\keyword{internal}
