\name{pharmavisR-package}
\alias{pharmavisR-package}
\alias{pharmavisR}
\docType{package}
\title{
\packageTitle{pharmavisR}
}
\description{
\packageDescription{pharmavisR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pharmavisR}
\packageIndices{pharmavisR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pharmavisR}

Maintainer: \packageMaintainer{pharmavisR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
