% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.default}
\alias{plot.survfit}
\title{Create a \code{ggplot} directly from an object through an S3 method}
\usage{
plot(x, ...)

\method{plot}{default}(x, ...)

\method{plot}{survfit}(
  survfit_object = NULL,
  y_label = NULL,
  x_label = NULL,
  x_units = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  fun = "surv",
  legend_position = "right"
)
}
\arguments{
\item{x}{object to be passed on to the method}

\item{...}{other arguments passed on to the method}

\item{survfit_object}{Object of class \code{survfit}}

\item{y_label}{\code{character} Label for the y-axis. When not specified, the default will do a proposal, depending on the \code{fun} argument.}

\item{x_label}{\code{character} Label for the x-asis. When not specified, the algorithm will look for "PARAM" information inside the list structure of the \code{survfit} object.
Note that this information is automatically added when using visR::vr_KM_est and when the input data has the variable "PARAM". If no "PARAM" information is available
"time" is used as label.}

\item{x_units}{Unit to be added to the x_label (x_label (x_unit)). Default is NULL.}

\item{x_ticks}{Ticks for the x-axis. When not specified, the default will do a proposal.}

\item{y_ticks}{Ticks for the y-axis. When not specified, the default will do a proposal based on the \code{fun} argument.}

\item{legend_position}{Specifies the legend position in the plot. Character values allowed are "top" "left" "bottom" "right". Numeric coordinates are also allowed.
Default is "right".}
}
\value{
Object of class \code{ggplot}  \code{ggsurvplot}.
}
\description{
S3 method for creating plots directly from objects using \code{ggplot2}, similar to base plot function.
The default method is base::plot.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

survfit_object <- KM_est(data = adtte, strata = "TRTP")

## Plot survival probability
plot(survfit_object = survfit_object, fun = "surv")
plot(survfit_object, fun = "pct")

## Plot cumulative hazard
plot(survfit_object, fun = "cloglog")
 
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
