% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.survfit_visr}
\alias{plot.survfit_visr}
\title{Plot a visR survfit object}
\usage{
\method{plot}{survfit_visr}(
  x = NULL,
  ...,
  y_label = NULL,
  x_label = NULL,
  x_units = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  fun = "surv",
  legend_position = "right"
)
}
\arguments{
\item{x}{Object of class \code{survfit_visr}}

\item{...}{other arguments passed on to the method}

\item{y_label}{\code{character} Label for the y-axis. When not specified, the default will do a proposal, depending on the \code{fun} argument.}

\item{x_label}{\code{character} Label for the x-asis. When not specified, the algorithm will look for "PARAM" information inside the list structure of the \code{survfit} object.
Note that this information is automatically added when using visR::estimate_KM and when the input data has the variable "PARAM". If no "PARAM" information is available
"time" is used as label.}

\item{x_units}{Unit to be added to the x_label (x_label (x_unit)). Default is NULL.}

\item{x_ticks}{Ticks for the x-axis. When not specified, the default will do a proposal.}

\item{y_ticks}{Ticks for the y-axis. When not specified, the default will do a proposal based on the \code{fun} argument.}

\item{fun}{TODO description}

\item{legend_position}{Specifies the legend position in the plot. Character values allowed are "top" "left" "bottom" "right". Numeric coordinates are also allowed.
Default is "right".}
}
\value{
Object of class \code{ggplot}  \code{ggsurvplot}.
}
\description{
Plot a visR survfit object
}
\examples{
library(survival)
library(visR)

# fit KM 
km_fit <- survival::survfit(Surv(AVAL, 1-CNSR) ~ TRTP, data=adtte)

# plot curves using base plot function
survival:::plot.survfit(km_fit)

# plot same curves using visR plot function

# estimate KM using visR wrapper
survfit_object <- visR::estimate_KM(data = adtte, strata = "TRTP")

## Plot survival probability
plot(survfit_object, fun = "surv")
plot(survfit_object, fun = "pct")

## Plot cumulative hazard
plot(survfit_object, fun = "cloglog")
 
}
