% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvr_attrition.R
\name{pvr_attrition}
\alias{pvr_attrition}
\title{Attrition}
\usage{
pvr_attrition(N_array, descriptions, complement_descriptions)
}
\arguments{
\item{N_array}{An array of integers of length M, where M is the number of 
steps in your cohort selection.}

\item{descriptions}{An array of characters of length M describing each 
step of the cohort selection (e.g. patients at least 18 years old at 
diagnosis).}

\item{complement_descriptions}{An array of characters of length M - 1 
describing the complement of each step in the cohort selection except for 
the first (e.g. patients less than 18 years old at diagnosis).}
}
\description{
Simple attrition diagrams for cohort selection
}
