% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{strata}
\alias{frailty}
\alias{cluster}
\alias{ridge}
\alias{pspline}
\alias{Surv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{survival}{\code{\link[survival]{cluster}}, \code{\link[survival]{frailty}}, \code{\link[survival]{pspline}}, \code{\link[survival]{ridge}}, \code{\link[survival]{strata}}, \code{\link[survival]{Surv}}}
}}

