% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_datatable}
\alias{render_datatable}
\title{Experimental internal function to help render a data.frame, risktable or tableone object as a datatable}
\usage{
render_datatable(data, title, download_format, source_cap)
}
\arguments{
\item{data}{Input data.frame or tibble to visualize}

\item{title}{Specify the title as a text string to be displayed in the rendered table.
Default is no title.}

\item{download_format}{Options formats generated for downloading the data.
Default is a list "c('copy', 'csv', 'excel')".'}

\item{source_cap}{String automatically compiled of data source and captions.}
}
\value{
A table data structure with possible interactive functionality depending on the choice of the engine.
}
\description{
Render a previously created datatable to html, rtf or latex
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
