% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_gt}
\alias{render_gt}
\title{Experimental function to render to a gt table.}
\usage{
render_gt(data, title, datasource, footnote)
}
\arguments{
\item{data}{Input data.frame or tibble to visualize}

\item{title}{Specify the title as a text string to be displayed in the rendered table.
Default is no title.}

\item{datasource}{String specifying the data source underlying the data set.
Default is no title.}

\item{footnote}{String specifying additional information to be displayed as a footnote
alongside the data source and specifications of statistical tests.}
}
\value{
A gt object.
}
\description{
Render a previously created datatable to gt
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
