% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{replace_hex_alpha}
\alias{replace_hex_alpha}
\title{Replaces the AA part of a #RRGGBBAA hex-colour.}
\usage{
replace_hex_alpha(colour, new_alpha)
}
\arguments{
\item{colour}{A character string of the format #RRGGBBAA.}

\item{new_alpha}{A two-character string with the new alpha value.}
}
\value{
A hex-encoded RGBA colour.
}
\description{
The function dissects a string of the format #RRGGBBAA into its components and replaces the alpha (AA) part.
}
\keyword{internal}
