% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_visR.R
\name{style_visR}
\alias{style_visR}
\title{Add visR theme to ggplot chart}
\usage{
style_visR(gg)
}
\arguments{
\item{gg}{ggplot object to style}
}
\value{

}
\description{
Add visR theme to ggplot chart
}
\examples{

library(ggplot2)

}
\keyword{visR}
