% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize.default}
\alias{summarize.default}
\title{Create variable summary for all other variable types}
\usage{
\method{summarize}{default}(x)
}
\arguments{
\item{x}{an object of any other class}
}
\description{
Create variable summary for all other variable types
}
