% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize.factor}
\alias{summarize.factor}
\title{Create variable summary for factors}
\usage{
\method{summarize}{factor}(x)
}
\arguments{
\item{x}{an object of class "factor"}
}
\description{
Create variable summary for factors
}
