% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize.integer}
\alias{summarize.integer}
\title{Create variable summary for numeric variables}
\usage{
\method{summarize}{integer}(x)
}
\arguments{
\item{x}{an object of class "integer"}
}
\description{
Create variable summary for numeric variables
}
