% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize.numeric}
\alias{summarize.numeric}
\title{Create variable summary for numeric variables}
\usage{
\method{summarize}{numeric}(x)
}
\arguments{
\item{x}{an object of class "numeric"}
}
\description{
Create variable summary for numeric variables
}
