% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize}
\alias{summarize}
\title{Calculate summary statistics for a vector}
\usage{
summarize(x)
}
\arguments{
\item{x}{an object}
}
\description{
Calculates several summary statistics for a vector depending on the vector class
}
