% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_long.default}
\alias{summarize_long.default}
\title{Create variable summary for all other variable types}
\usage{
\method{summarize_long}{default}(x)
}
\arguments{
\item{x}{an object of any other class}
}
\value{
List of counts for unique and missing values in \code{x}.
}
\description{
Create variable summary for all other variable types
}
