% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_long.integer}
\alias{summarize_long.integer}
\title{Create variable summary for numeric variables}
\usage{
\method{summarize_long}{integer}(x)
}
\arguments{
\item{x}{an object of class "integer"}
}
\description{
Create variable summary for numeric variables
}
