% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_short.factor}
\alias{summarize_short.factor}
\title{Create variable summary for factors}
\usage{
\method{summarize_short}{factor}(x)
}
\arguments{
\item{x}{an object of class "factor"}
}
\value{
Short list of summary statistics for the input factors.
}
\description{
Calculates N and \% of occurrence for each factor value
}
