% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_short}
\alias{summarize_short}
\title{Create abbreviated variable summary for table1}
\usage{
summarize_short(x)
}
\arguments{
\item{x}{a vector to be summarized}
}
\value{
A summarized less detailed version of the input.
}
\description{
This function creates summaries combines multiple summary measures in a single formatted string.
}
