% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_tab1.factor}
\alias{summarize_tab1.factor}
\title{Create variable summary for factors}
\usage{
\method{summarize_tab1}{factor}(x)
}
\arguments{
\item{x}{an object of class "factor"}
}
\description{
Calculates N and \% of occurence for each factor value
}
