% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_tab1.integer}
\alias{summarize_tab1.integer}
\title{Create variable summary for integer variables}
\usage{
\method{summarize_tab1}{integer}(x)
}
\arguments{
\item{x}{an object of class "integer"}
}
\description{
Calculates mean (standard deviation), median (IQR), min-max range and N/\% missing elements
for a integer vector.
}
