% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_tab1.numeric}
\alias{summarize_tab1.numeric}
\title{Create variable summary for numeric variables}
\usage{
\method{summarize_tab1}{numeric}(x)
}
\arguments{
\item{x}{an object of class "numeric"}
}
\description{
Calculates mean (standard deviation), median (IQR), min-max range and N/\% missing elements
for a numeric vector.
}
