% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_tab1}
\alias{summarize_tab1}
\title{Create abbreviated variable summary for table1}
\usage{
summarize_tab1(x)
}
\arguments{
\item{x}{a vector to be summarized}
}
\description{
This function creates summaries combines multiple summary measures in a single formatted string.
}
