% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_CDISC.R
\name{Surv_CDISC}
\alias{Surv_CDISC}
\title{Create a Survival Object from CDISC Data}
\usage{
Surv_CDISC(AVAL, CNSR)
}
\arguments{
\item{AVAL}{The follow-up time. The follow-up time is assumed to originate from zero.
When no argument is passed, the default value is a column/vector named \code{AVAL}}

\item{CNSR}{The censoring indicator where \code{1=censored} and \code{0=death/event}.
When no argument is passed, the default value is a column/vector named \code{CNSR}}
}
\value{
Object of class 'Surv'
}
\description{
Create a survival object, usually used as a response variable in a model
formula. The arguments are named after the CDISC data standard.
The \code{CNSR} argument expects a numeric column/vector that indicates
an observation is censored.

The \code{AVAL} and \code{CNSR} arguments are passed to
\code{survival::Surv(time = AVAL, event = 1 - CNSR, type = 'right', origin = 0)}.
}
\examples{
library(survival)

survfit(Surv_CDISC() ~ SEX, data = adtte)
adtte \%>\%
  estimate_KM(formula = Surv_CDISC() ~ SEX)

# When using CDSIC data, you can specify arguments or rely on defaults
with(adtte, Surv_CDISC(AVAL, CNSR)) \%>\% head()
with(adtte, Surv_CDISC()) \%>\% head()
}
\seealso{
\code{\link[survival:Surv]{survival::Surv()}}, \code{\link[=estimate_KM]{estimate_KM()}}
}
