% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_caption}
\alias{table_caption}
\title{Create a caption for any table}
\usage{
table_caption(
  caption,
  label = knitr::opts_current$get("label"),
  width = knitr::opts_current$get("out.width")
)
}
\arguments{
\item{caption}{The table caption}

\item{label}{Label}

\item{width}{Width}
}
\description{
Create a caption for any table
}
