% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{the_lhs}
\alias{the_lhs}
\title{Find the "lhs" symbol in the pipeline}
\usage{
the_lhs()
}
\value{
Left-hand sided symbol as string in the magrittr pipe.
}
\description{
This function finds the left-hand sided symbol in a magrittr pipe.
}
\examples{
library(magrittr)

## example
blah <- function(x) the_lhs()
adtte \%>\%
  blah()
}
\references{
\url{https://github.com/tidyverse/magrittr/issues/115#issuecomment-173894787}
}
\author{
Steven Haesendonckx {shaesen2@its.jnj.com}
}
