% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visR-package.R
\docType{package}
\name{visR-package}
\alias{visR}
\alias{visR-package}
\title{visR: Clinical Graphs and Tables Adhering to Graphical Principles}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

To enable fit-for-purpose, reusable clinical and medical research focused visualizations and tables with sensible defaults and based on graphical principles as described in: "Vandemeulebroecke et al. (2018)" \doi{10.1002/pst.1912}, "Vandemeulebroecke et al. (2019)" \doi{10.1002/psp4.12455}, and "Morris et al. (2019)" \doi{10.1136/bmjopen-2019-030215}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openpharma.github.io/visR/}
  \item \url{https://github.com/openpharma/visR}
  \item Report bugs at \url{https://github.com/openpharma/visR/issues}
}

}
\author{
\strong{Maintainer}: Mark Baillie \email{bailliem@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Diego Saldana \email{diego.saldana@roche.com}
  \item Charlotta Fruechtenicht \email{charlotta.fruechtenicht@roche.com}
  \item Marc Vandemeulebroecke \email{marc.vandemeulebroecke@novartis.com}
  \item Thanos Siadimas \email{thanos.siadimas@roche.com}
  \item Pawel Kawski \email{pawel.kawski@contractors.roche.com}
  \item Steven Haesendonckx \email{shaesen2@its.jnj.com}
  \item James Black \email{james.black.jb2@roche.com}
  \item Pelagia Alexandra Papadopoulou \email{PPapadop@its.jnj.com}
  \item Tim Treis \email{tim.treis@outlook.de}
  \item Rebecca Albrecht \email{rebecca.albrecht.dietsch@gmail.com}
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
}

Other contributors:
\itemize{
  \item Ardalan Mirshani \email{ardeeshany@gmail.com} [contributor]
}

}
\keyword{internal}
