% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visR.R
\docType{package}
\name{visR}
\alias{visR}
\title{\code{visR} package}
\description{
visR
}
\details{
See the README on
\href{https://github.com/openpharma/visR}{GitHub}

For "visible binding note solution" see here \href{https://github.com/STAT545-UBC/Discussion/issues/451}
}
