% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr_plot_survival.R
\name{visr_plot_risk}
\alias{visr_plot_risk}
\title{Create a ggplot of a risk table}
\usage{
visr_plot_risk(x, timeby = 10)
}
\arguments{
\item{x}{survfit object}

\item{timeby}{Integer to indicate range to summarise risk table}
}
\value{
ggplot object
}
\description{
Create a ggplot of a risk table
}
\examples{
library(survival)
library(magrittr)
library(dplyr)
library(ggplot2)
library(broom)
fit <- survival::survfit(Surv(time, status) ~ sex, data = lung)
visr_plot_risk(fit, 50) \%>\%
ggplot(aes(x = time, y = strata_id, label = n.risk, group= strata)) + 
geom_text() 
}
