% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr_plot_survival.R
\name{visr_plot_surv}
\alias{visr_plot_surv}
\title{Plot step lines based on a survival or time to event analysis}
\usage{
visr_plot_surv(x)
}
\arguments{
\item{x}{tidy tibble dataset. Based on broom tidied dataset.}
}
\value{
ggplot object
}
\description{
The function assumes a tidied tibble dataset with the following variables:
time 
estimate
strata
}
\examples{
library(survival)
library(magrittr)
library(dplyr)
library(ggplot2)
library(broom)
survival::survfit(Surv(time, status) ~ sex, data = lung)\%>\%
  tidy() \%>\%
  visr_plot_surv()
}
