% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr_plot_survival.R
\name{visr_summary_surv}
\alias{visr_summary_surv}
\title{Function to summarise risk by interval
See code from here https://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/}
\usage{
visr_summary_surv(x, timeby)
}
\arguments{
\item{x}{surv object to summarise}

\item{timeby}{intervals to disaply the risk table}
}
\value{
tidy tibble
}
\description{
Function to summarise risk by interval
See code from here https://mcfromnz.wordpress.com/2012/05/05/kaplan-meier-survival-plot-with-at-risk-table-by-sub-groups/
}
\examples{
library(survival)
library(magrittr)
library(dplyr)
library(ggplot2)
library(broom)
fit <- survival::survfit(Surv(time, status) ~ sex, data = lung)
visr_summary_surv(fit, 50)
}
