% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_attrition.R
\name{vr_attrition}
\alias{vr_attrition}
\title{Attrition}
\usage{
vr_attrition(
  N_array,
  descriptions,
  complement_descriptions,
  output_path = NULL,
  display = FALSE,
  wrap_width = 50,
  node_width = 3,
  font_size = 8
)
}
\arguments{
\item{N_array}{An array of integers of length M, where M is the number of
steps in your cohort selection.}

\item{descriptions}{An array of characters of length M describing each
step of the cohort selection (e.g. patients at least 18 years old at
diagnosis).}

\item{complement_descriptions}{An array of characters of length M - 1
describing the complement of each step in the cohort selection except for
the first (e.g. patients less than 18 years old at diagnosis).}

\item{output_path}{The path to the output (either a .ps or a .svg file).}

\item{display}{Whether to display the dot string.}

\item{wrap_width}{The width of the text for text wrapping, as used in
the strwrap function..}

\item{node_width}{The width of the node as used in graphviz (dot).}

\item{font_size}{The label font size as used in graphviz (dot).}
}
\description{
Simple attrition diagrams for cohort selection
}
