% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_est_kaplan_meier.R
\name{vr_est_kaplan_meier}
\alias{vr_est_kaplan_meier}
\title{Create Tidy Survival Object}
\usage{
vr_est_kaplan_meier(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Create Tidy Survival Object
}
