% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_est_km_risk_table.R
\name{vr_est_km_risk_table}
\alias{vr_est_km_risk_table}
\title{Create Tidy Risk Table}
\usage{
vr_est_km_risk_table(data, min_at_risk = 3)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Create Tidy Risk Table
}
