% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_kaplan_meier.R
\name{vr_kaplan_meier}
\alias{vr_kaplan_meier}
\title{Create Kaplan-Meier Curve for Survival Data}
\usage{
vr_kaplan_meier(
  data,
  title = "",
  abbreviations = NULL,
  variable_definitions = NULL,
  N = NULL,
  N_unit = "patients",
  time_unit = "days",
  data_source = NULL
)
}
\arguments{
\item{data}{}

\item{title}{}

\item{abbreviations}{}

\item{variable_definitions}{}

\item{N}{}

\item{N_unit}{}

\item{time_unit}{}

\item{data_source}{}
}
\value{

}
\description{
Create Kaplan-Meier Curve for Survival Data
}
