% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_kaplan_meier_summary.R
\name{vr_kaplan_meier_summary}
\alias{vr_kaplan_meier_summary}
\title{Plot Kaplan-Meier Summary Table for Existing}
\usage{
vr_kaplan_meier_summary(data, equation)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Plot Kaplan-Meier Summary Table for Existing
}
