% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_p_val.R
\name{vr_p_val.factor}
\alias{vr_p_val.factor}
\title{P value test for factors variables}
\usage{
\method{vr_p_val}{factor}(x, group)
}
\arguments{
\item{x}{Vector for variable of interest}

\item{group}{Vector with grouping variable data}
}
\description{
P value test for factors variables
}
