% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_p_val.R
\name{vr_p_val}
\alias{vr_p_val}
\title{Adds p-values to tableone}
\usage{
vr_p_val(x, ...)
}
\arguments{
\item{x}{Vector for variable of interest}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Adds p-values to tableone
}
