% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plot.R
\name{vr_plot.survfit}
\alias{vr_plot.survfit}
\title{Plot Kaplan-Meier Curve for Existing Tidy Survival Object}
\usage{
\method{vr_plot}{survfit}(
  survfit_object = NULL,
  y_label = NULL,
  x_label = NULL,
  x_units = NULL,
  time_ticks = NULL,
  y_ticks = NULL,
  fun = "surv",
  legend_position = "right",
  debug = F
)
}
\value{
ggplot object
}
\description{
TODO: Currently we choose between survival and cumhazard using the cumhazard argument => use funs = argument to define what to plot
Apply the fun argument similar to survival:::plot.survfit 
an arbitrary function defining a transformation of the survival curve. 
fun=log => axis labeled with log(S) values 
fun=sqrt => square root scale. 
"S" gives the usual survival curve,
"log" is the same as using the log=T option
"event" or "F" plots the empirical CDF F(t)= 1-S(t) (f(y) = 1-y),
"cumhaz" plots the cumulative hazard function (see details)
"cloglog" creates a complimentary log-log survival plot (f(y) = log(-log(y)) along with log scale for the x-axis).
The terms "identity" and "surv" are allowed as synonyms for type="S".
}
\details{
TODO: legend_pos option where: inside vs outside. This determines the ggrob later on
}
\examples{
# TODO: Define an example for this function
library(survival)
library(glue)
library(dplyr)
library(tidyr)
library(ggplot2)

fit <- vr_KM_est(data = adtte, strata = "TRTP")

## Plot survival probability
vr_plot(survfit_object = fit)

## Plot cumulative hazard
vr_plot(survfit_object = fit, fun = "cumhaz", debug=T)
}
