% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plot_forest.R
\name{vr_plot_forest}
\alias{vr_plot_forest}
\title{Render a simple forest plot given a tibble}
\usage{
vr_plot_forest(td)
}
\arguments{
\item{td}{tidied tibble}
}
\value{
ggplot object
}
\description{
Render a simple forest plot given a tibble
}
\examples{
library(RBesT) 
library(dplyr)
library(ggplot2) 

example(crohn) 
vr_tidy_rbest(map_crohn)
map_crohn \%>\% vr_tidy_rbest() \%>\% filter(model == "meta") \%>\%
vr_plot_forest()

map_crohn \%>\% vr_tidy_rbest() \%>\% filter(model == "stratified") \%>\%
vr_plot_forest()

map_crohn \%>\% vr_tidy_rbest() \%>\% vr_plot_forest() + facet_wrap(~ model)

}
