% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plt_kaplan_meier.R
\name{vr_plt_kaplan_meier}
\alias{vr_plt_kaplan_meier}
\title{Plot Kaplan-Meier Curve for Existing Tidy Survival Object}
\usage{
vr_plt_kaplan_meier(
  data,
  title = "",
  abbreviations = "",
  variable_definitions = "",
  N,
  N_unit = "patients"
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Plot Kaplan-Meier Curve for Existing Tidy Survival Object
}
