% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plt_km_risk_table.R
\name{vr_plt_km_risk_table}
\alias{vr_plt_km_risk_table}
\title{Plot Kaplan-Meier Risk Table for Existing Tidy Risk Object}
\usage{
vr_plt_km_risk_table(risk_table_data, time_unit)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Plot Kaplan-Meier Risk Table for Existing Tidy Risk Object
}
