% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_render_table.R
\name{vr_render_table}
\alias{vr_render_table}
\title{Render a dataframe or tibble}
\usage{
vr_render_table(
  data,
  title,
  caption,
  datasource,
  output_format = "html",
  engine = "gt"
)
}
\arguments{
\item{data}{The dataframe or tibble to visualise}

\item{title}{Table title to include in the rendered table}

\item{caption}{Table caption to include in the rendered table}

\item{datasource}{String specifying the datasource underlying the data set}

\item{output_format}{If TRUE, the summary statistics for the overall dataset are also calculated}

\item{engine}{If html is selected as output format, one can chose between using kable, gt and DT as engine to create the output table}
}
\description{
Render a previously created dataframe or tibble to html, rtf or latex

Create formatted table
}
