% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize.default}
\alias{vr_summarize.default}
\title{Create variable summary for all other variable types}
\usage{
\method{vr_summarize}{default}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Create variable summary for all other variable types
}
