% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize.factor}
\alias{vr_summarize.factor}
\title{Create variable summary for factors}
\usage{
\method{vr_summarize}{factor}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Create variable summary for factors
}
