% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize}
\alias{vr_summarize}
\title{Calculate summary statistics for a vector}
\usage{
vr_summarize(x)
}
\arguments{
\item{x}{an object}
}
\description{
Calculates several summary statistics for a vector depending on the vector class
}
