% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize_tab1}
\alias{vr_summarize_tab1}
\title{Create variable summary for table1}
\usage{
vr_summarize_tab1(x)
}
\arguments{
\item{x}{placeholder}
}
\description{
Create variable summary for table1
}
