% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_tidy_rbest.R
\name{vr_tidy_rbest}
\alias{vr_tidy_rbest}
\title{Returns a tidied RBesT gMAP object}
\usage{
vr_tidy_rbest(x, prob = 0.95)
}
\arguments{
\item{x}{RBesT gMAP object}

\item{prob}{probability range for uncertainty interval}
}
\value{
td tidied tibble
}
\description{
Returns a tidied RBesT gMAP object
}
\examples{
library(RBesT)
example(crohn)
vr_tidy_rbest(map_crohn)

map_crohn \%>\% vr_tidy_rbest() \%>\% filter(model == "meta") \%>\% ggplot(aes( x =
study, y = estimate, ymin = conf.low, ymax = conf.high )) +
geom_pointrange(show.legend = FALSE, width = 1) + coord_flip()

}
