% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_var_label.R
\name{vr_var_label}
\alias{vr_var_label}
\title{Sets the label attribute on a data frame}
\usage{
vr_var_label(dat, labels)
}
\arguments{
\item{dat}{Data frame}

\item{labels}{A named list, with the name being the variable and the
value being the label to apply.}
}
\value{
Returns the data frame after applying the labels
}
\description{
A helper function to apply a set of attribute labels to variables
in a data frame. This is useful if you will be passing your data frame
to vr_create_tableone because the labels will be used instead of the
variable names.
}
\examples{
mtcars_labeled <- vr_var_label(
    mtcars,
    labels = list(
        "mpg" = "Miles Per Gallon",
        "cyl" = "Number Of Cylinders"
    )
)
}
