% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\alias{add_CI.ggsurvfit}
\title{Add pointwise Confidence Interval to a plot created through visR using S3 method}
\usage{
add_CI(gg, ...)

\method{add_CI}{ggsurvfit}(gg, alpha = 0.1, style = "ribbon", linetype = 2, ...)
}
\arguments{
\item{gg}{A plot created with visR}

\item{...}{other arguments passed on to the method}

\item{x}{visR plot}
}
\value{
Pointwise confidence interval overlayed on a visR plot
}
\description{
S3 method for adding a pointwise confidence interval to a visR plot.
    No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
vr_plot(survfit_object) \%>\%
  add_CI(alpha = 0.1, style = "step", linetype = 3)
 
}
\author{
Steven Haesendonckx {shaesen2@its.jnj.com}
}
