% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CI.R
\name{add_CI}
\alias{add_CI}
\alias{add_CI.ggsurvfit}
\title{Add pointwise Confidence Interval to a an object created by visR through an S3 method}
\usage{
add_CI(gg, ...)

\method{add_CI}{ggsurvfit}(gg, alpha = 0.1, style = "ribbon", linetype = 2, ...)
}
\arguments{
\item{gg}{A ggplot created with visR}

\item{...}{other arguments passed on to the method to modify \code{\link[ggplot2]{geom_ribbon}}}

\item{alpha}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is 0.1.}

\item{style}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is "ribbon".}

\item{linetype}{aesthetic of ggplot2 \code{\link[ggplot2]{geom_ribbon}}. Default is 2.}
}
\value{
Pointwise confidence interval overlayed on a visR ggplot
}
\description{
S3 method for adding a pointwise confidence interval to a object created with visR.
No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
vr_plot(survfit_object) \%>\%
  add_CI(alpha = 0.1, style = "step", linetype = 3)

}
\author{
Steven Haesendonckx
}
