% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_highlight.R
\name{add_highlight}
\alias{add_highlight}
\alias{add_highlight.ggsurvfit}
\title{Highlight a specific strata}
\usage{
add_highlight(gg, ...)

\method{add_highlight}{ggsurvfit}(gg = NULL, strata = NULL, bg_alpha_multiplier = 0.2, ...)
}
\arguments{
\item{gg}{A ggsurvfit, ideatlly created with visR}

\item{...}{other arguments passed on to the method}

\item{strata}{String representing the name and value of the strata to be highlighted as shown in the legend.}

\item{bg_alpha_multiplier}{Factor with which the \code{alpha} values of all but the specified strata will be multiplied.}
}
\value{
The input \code{ggsurvfit} object with adjusted \code{alpha} values
}
\description{
S3 method for highlighting a specific strata by lowering the opacity of all other strata.
}
\examples{

adtte \%>\%
  visR::estimate_KM(strata = "SEX") \%>\%
  visR::visr() \%>\%
  visR::add_CI(alpha = 0.4) \%>\%
  visR::add_highlight(strata = "SEX=M", bg_alpha_multiplier = 0.2)

strata = c("TRTP=Placebo", "TRTP=Xanomeline Low Dose")

adtte \%>\%
  visR::estimate_KM(strata = "TRTP") \%>\%
  visR::visr() \%>\%
  visR::add_CI(alpha = 0.4) \%>\%
  visR::add_highlight(strata = strata, bg_alpha_multiplier = 0.2)

}
