% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable.R
\name{add_risktable}
\alias{add_risktable}
\alias{add_risktable.ggsurvfit}
\title{Add risk tables to visR plots through an S3 method}
\usage{
add_risktable(gg, ...)

\method{add_risktable}{ggsurvfit}(gg, risktable)
}
\arguments{
\item{gg}{visR plot of class \code{ggsurvfit}}

\item{...}{other arguments passed on to the method}

\item{risktable}{A risktable created with method 'get_risktable'}
}
\value{
Object of class \code{ggplot} with added risk table.
}
\description{
S3 method for adding risk tables to visR plots.
No default method is available at the moment.
}
\examples{
\donttest{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)
library(gtable)

## Display 2 risk tables, 1 per statlist
adtte \%>\%
  vr_KM_est(strata = "TRTP") \%>\%
  vr_plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,group = "statlist"
                )
                
## Display 2 risk tables, 1 per stratum
adtte \%>\%
  vr_KM_est(strata = "TRTP") \%>\%
  vr_plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,group = "strata"
                )

## Display overall risk table
adtte \%>\%
  vr_KM_est(strata = "TRTP") \%>\%
  vr_plot() \%>\%
  add_risktable( min_at_risk = 3
                ,label = c("Subjects at Risk", "Censored")
                ,statlist = c("n.risk", "n.censor")
                ,collapse = TRUE
                )

}

}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
