% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable.R
\name{add_risktable}
\alias{add_risktable}
\alias{add_risktable.ggsurvfit}
\title{Add risk tables to visR plots through an S3 method}
\usage{
add_risktable(gg, ...)

\method{add_risktable}{ggsurvfit}(
  gg,
  min_at_risk = 0,
  times = NULL,
  statlist = c("n.risk"),
  label = "At risk",
  group = "strata",
  collapse = FALSE
)
}
\arguments{
\item{gg}{visR plot of class \code{ggsurvfit}}

\item{...}{other arguments passed on to the method add_risktable}

\item{min_at_risk}{\code{numeric} The cutoff for number of participants at risk to display. This minimum is applied across strata.
Default is 0.}

\item{times}{Numeric vector indicating the times at which the risk set, censored subjects, events are calculated.}

\item{statlist}{Character vector indicating which summary data to present. Current choices are "n.risk" "n.event" "n.censor".
Default is "n.risk".}

\item{label}{Character vector with labels for the statlist. Default matches "n.risk" with "At risk", "n.event" with "Events" and "n.censor"
with "Censored".}

\item{group}{String indicating the grouping variable for the risk tables.
Current options are:
\itemize{
\item{"strata": groups the risk tables per stratum.
The \code{label} specifies the label within each risk table. The strata levels
are used for the titles of the risk tables. This is the default}
\item{"statlist": groups the risk tables per statlist.
The \code{label} specifies the title for each risk tabel. The strata levels
are used for labeling within each risk table.}
}
Default is "strata".}

\item{collapse}{Boolean, indicates whether to present the data overall.
Default is FALSE.}
}
\value{
Object of class \code{ggplot} with added risk table.
}
\description{
S3 method for adding risk tables to visR plots. The function has following workflow:
\itemize{
\item{
The risktables are calculated using \code{\link[visR]{get_risktable}}
}
\item{
The risktables are placed underneath visR plots using \code{\link[cowplot]{plot_grid}}
}
\item{
Both the initial visR plot as the individual risktables are stored as attribute \code{component}
in the final object to allow post-modification of the individual plots if desired
}
}
}
\examples{

## Display 2 risk tables, 1 per statlist
adtte \%>\%
  visR::estimate_KM(strata = "TRTP") \%>\%
  visR::visr() \%>\%
  visR::add_risktable( min_at_risk = 3
                      ,label = c("Subjects at Risk", "Censored")
                      ,statlist = c("n.risk", "n.censor")
                      ,group = "statlist"
                     )
                
## Display 2 risk tables, 1 per stratum
adtte \%>\%
  visR::estimate_KM(strata = "TRTP") \%>\%
  visR::visr() \%>\%
  visR::add_risktable( min_at_risk = 3
                      ,label = c("Subjects at Risk", "Censored")
                      ,statlist = c("n.risk", "n.censor")
                      ,group = "strata"
                     )

## Display overall risk table
adtte \%>\%
  visR::estimate_KM(strata = "TRTP") \%>\%
  visR::visr() \%>\%
  visR::add_risktable( min_at_risk = 3
                      ,label = c("Subjects at Risk", "Censored")
                      ,statlist = c("n.risk", "n.censor")
                      ,collapse = TRUE
                     )

}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
