% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable.R
\name{add_risktable}
\alias{add_risktable}
\alias{add_risktable.ggsurvfit}
\title{Add risk tables to visR plots through an S3 method}
\usage{
add_risktable(gg, ...)

\method{add_risktable}{ggsurvfit}(
  gg,
  min_at_risk = 0,
  time_ticks = NULL,
  display = c("n.risk"),
  title = c("Subjects at risk")
)
}
\arguments{
\item{gg}{visR plot of class \code{ggsurvfit}}

\item{...}{other arguments passed on to the method}

\item{min_at_risk}{\code{numeric} The cutoff for number of subjects to display. Default is 0.}

\item{time_ticks}{Numeric vector with the points along the x-axis at which the summary data needs to be provided.}

\item{display}{Character vector indicating which summary data to present. Current choices are "n.risk" "n.event" "n.censor".}

\item{title}{Character vector with titles for the summary tables.}
}
\value{
Object of class \code{ggplot} with added risk table.
}
\description{
S3 method for adding risk tables to visR plots.
No default method is available at the moment.
}
\examples{
\donttest{
library(survival)
library(dplyr)
library(tidyr)
library(ggplot2)
library(cowplot)
library(gtable)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
vr_plot(survfit_object) \%>\%
  add_risktable(min_at_risk = 3, title = c("blah"), display = c("n.risk", "n.censor"))
  }
 
}
\seealso{
\code{\link[cowplot]{plot_grid}}
}
\author{
Steven Haesendonckx
}
