% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cuminc.R
\name{estimate_cuminc}
\alias{estimate_cuminc}
\title{Competing Events Cumulative Incidence}
\usage{
estimate_cuminc(
  data,
  strata = NULL,
  CNSR = "CNSR",
  AVAL = "AVAL",
  conf.int = 0.95,
  ...
)
}
\arguments{
\item{data}{The name of the dataset for Time-to-Event analysis based on the Analysis Data Model (ADaM) principles. The dataset is expected to have
one record per subject per analysis parameter. Rows in which the analysis variable (AVAL) or the censor variable (CNSR) contain NA, are removed during analysis.}

\item{strata}{Character vector, representing the strata for Time-to-Event analysis. When NULL, an overall analysis is performed.
Default is NULL.}

\item{CNSR}{Column name indicating the outcome and censoring statuses.
Column must be a factor and the first level indicates censoring, the
next level is the outcome of interest, and the remaining levels are the
competing events. Default is \code{"CNSR"}}

\item{AVAL}{Analysis value for Time-to-Event analysis. Default is "AVAL", as per CDISC ADaM guiding principles.}

\item{conf.int}{Confidence internal level. Default is 0.95.}

\item{...}{Additional argument passed to \code{tidycmprsk::cuminc()}}
}
\value{
A cumulative incidence object as explained at https://mskcc-epi-bio.github.io/tidycmprsk/reference/cuminc.html
}
\description{
Function creates a cumulative incidence object using the
\code{tidycmprsk::cuminc()} function.
}
\examples{
estimate_cuminc(
  tidycmprsk::trial,
  strata = "trt",
  CNSR = "death_cr",
  AVAL = "ttdeath"
) \%>\%
  visr() \%>\%
  add_CI() \%>\%
  add_risktable(statlist = c("n.risk", "cum.event"))
}
