% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{get_alpha_from_hex_colour}
\alias{get_alpha_from_hex_colour}
\title{Extract the numerical alpha representation of #RRGGBBAA colour}
\usage{
get_alpha_from_hex_colour(hex_colour)
}
\arguments{
\item{hex_colour}{A string in the format '#RRGGBBAA'}
}
\value{
The numeric representation of the colors' alpha value, rounded to 2 digits.
}
\description{
This function dissects a #RRGGBBAA colour into its components and then returns the numeric representation of the alpha (AA) value.
}
\keyword{internal}
