% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue}
\alias{get_pvalue}
\alias{get_pvalue.survfit}
\title{Summarize the test for equality across strata from a survival object of class \code{survfit} using S3 method}
\usage{
get_pvalue(x, ...)

\method{get_pvalue}{survfit}(
  survfit_object,
  ptype = "All",
  rho = NULL,
  statlist = c("test", "Chisq", "df", "p"),
  ...
)
}
\arguments{
\item{x}{S3 object}

\item{...}{other arguments passed on to the method}

\item{survfit_object}{An object of class `survfit`}

\item{ptype}{Character vector containing the type of p-value desired. Current options are "Log-Rank" "Wilcoxon" "Tarone-Ware" "Custom" "All".
"Custom" allows the user to specify the weights on the Kaplan-Meier estimates using the argument `rho`.
The default is "All" displaying all types possible. When `rho` is specified in context of "All", also a custom p-value is displayed.}

\item{rho}{
a scalar parameter that controls the type of test.
}

\item{statlist}{Character vector containing the desired information to be displayed. The order of the arguments determines the order in which
they are displayed in the final result. Default is the test name ("test"), Chisquare test statistic ("Chisq"), degrees of freedom ("df") and
p-value ("p").}
}
\value{
A tibble with summary measures for the Test of Equality Across Strata
}
\description{
S3 method for extracting information regarding equality across strata.
    No default method is available at the moment.
}
\examples{
library(survival)
library(dplyr)
library(tidyr)

survfit_object <- survival::survfit(data = adtte, Surv(AVAL, 1-CNSR) ~ TRTP)
get_pvalue(surv_object)
 
}
\seealso{
\code{\link[survival]{survdiff}}
}
\author{
Steven Haesendonckx {shaesen2@its.jnj.com}
}
