% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risktable.R
\name{get_risktable}
\alias{get_risktable}
\alias{get_risktable.survfit}
\title{Obtain risk tables for tables and plots}
\usage{
get_risktable(
  survfit_object,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = "At risk",
  group = "strata",
  collapse = FALSE
)

\method{get_risktable}{survfit}(
  survfit_object,
  min_at_risk = 0,
  breaks = NULL,
  statlist = c("n.risk"),
  label = NULL,
  group = "strata",
  collapse = FALSE
)
}
\arguments{
\item{survfit_object}{an object of class \code{survfit}}

\item{min_at_risk}{\code{numeric} The cutoff for number of subjects to display. Default is 0.}

\item{breaks}{Single numeric or numeric vector indicating breaks.}

\item{statlist}{Character vector indicating which summary data to present. Current choices are "n.risk" "n.event" "n.censor".}

\item{label}{Character vector with labels for the statlist. Default matches "n.risk" with "At risk", "n.event" with "Events" and "n.censor"
with "Censored".}

\item{group}{String indicating the grouping variable for the risk tables. Current options are:
\itemize{
\item{"strata": groups the risk tables per stratum. The \code{label} specifies the lables used within each risk table. This is the default}
\item{"statlist": groups the risk tables per statlist. The \code{label} specifies the title for each risk tabel. The strata levels
are used for labeling within each risk table.}
} "strata" to group the risk tables
per strata, and "statlist" to group the risk tables}

\item{collapse}{Boolean, indicates whether to present the data overall, rather than per strata.
Default is FALSE.}
}
\description{
Create a risktable from a survival object to display in tables or plots
}
