% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_visr.R
\name{legendopts}
\alias{legendopts}
\title{Translates options for legend into a list that can be passed to ggplot2}
\usage{
legendopts(legend_position = "right", legend_orientation = NULL)
}
\arguments{
\item{legend_position}{Default = "right".}

\item{legend_orientation}{Default = NULL.}
}
\value{
List of legend options for ggplot2.
}
\description{
This function takes the legend position and orientation, defined by the user and puts them into a list for ggplot2.
}
