% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\alias{render.tableone}
\alias{render.risktable}
\alias{render.data.frame}
\title{Render a data.frame or tibble}
\usage{
render(
  data,
  title,
  datasource,
  footnote = "",
  output_format = "html",
  engine = "gt",
  download_format = c("copy", "csv", "excel")
)

\method{render}{tableone}(
  data,
  title,
  datasource,
  footnote = "",
  output_format = "html",
  engine = "gt",
  download_format = NULL
)

\method{render}{risktable}(
  data,
  title,
  datasource,
  footnote = "",
  output_format = "html",
  engine = "gt",
  download_format = NULL
)

\method{render}{data.frame}(
  data,
  title,
  datasource,
  footnote = "",
  output_format = "html",
  engine = "gt",
  download_format = c("copy", "csv", "excel")
)
}
\arguments{
\item{data}{The data.frame or tibble to visualize}

\item{title}{Table title to include in the rendered table}

\item{datasource}{String specifying the data source underlying the data set}

\item{footnote}{String specifying additional information to be displayed in the table note alongside the data source and specifications of statistical tests.}

\item{output_format}{Type of output that is returned, can be html or latex}

\item{engine}{If html is selected as output format, one can chose between
using kable, gt and DT as engine to create the output table}

\item{download_format}{Options generated for downloading the data}
}
\value{
A table-like data structure, possibly interactive depending on the choice of the engine

A table-like data structure, possibly interactive depending on the choice of the engine
}
\description{
Render a previously created data.frame to html,
rtf or latex
}
