% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_long.factor}
\alias{summarize_long.factor}
\title{Create variable summary for factors}
\usage{
\method{summarize_long}{factor}(x)
}
\arguments{
\item{x}{an object of class "factor"}
}
\description{
Create variable summary for factors
}
