% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_long.numeric}
\alias{summarize_long.numeric}
\title{Create variable summary for numeric variables}
\usage{
\method{summarize_long}{numeric}(x)
}
\arguments{
\item{x}{an object of class "numeric"}
}
\value{
numeric value
}
\description{
Create variable summary for numeric variables
}
