% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_long}
\alias{summarize_long}
\alias{summarize_long.factor}
\alias{summarize_long.integer}
\alias{summarize_long.numeric}
\alias{summarize_long.default}
\title{Calculate summary statistics from a vector}
\usage{
summarize_long(x)

\method{summarize_long}{factor}(x)

\method{summarize_long}{integer}(x)

\method{summarize_long}{numeric}(x)

\method{summarize_long}{default}(x)
}
\arguments{
\item{x}{an object}
}
\value{
A summarized version of the input.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Calculates several summary statistics. The summary statistics
depend on the vector class
}
