% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{summarize_short}
\alias{summarize_short}
\alias{summarize_short.factor}
\alias{summarize_short.numeric}
\alias{summarize_short.integer}
\alias{summarize_short.default}
\title{Create abbreviated variable summary for table1}
\usage{
summarize_short(x)

\method{summarize_short}{factor}(x)

\method{summarize_short}{numeric}(x)

\method{summarize_short}{integer}(x)

\method{summarize_short}{default}(x)
}
\arguments{
\item{x}{a vector to be summarized}
}
\value{
A summarized less detailed version of the input.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
This function creates summaries combines multiple summary
measures in a single formatted string. Create variable summary for numeric variables. Calculates mean
(standard deviation), median (IQR), min-max range and N/\% missing elements
for a numeric vector.

Create variable summary for integer variables
Calculates mean (standard deviation), median (IQR), min-max range
and N/\% missing elements for a integer vector.
}
