% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv_CDISC.R
\name{Surv_CDISC}
\alias{Surv_CDISC}
\title{Create a Survival Object from CDISC Data}
\usage{
Surv_CDISC(AVAL, CNSR)
}
\arguments{
\item{AVAL}{The follow-up time. The follow-up time is assumed to originate from zero.
When no argument is passed, the default value is a column/vector named \code{AVAL}}

\item{CNSR}{The censoring indicator where \code{1=censored} and \code{0=death/event}.
When no argument is passed, the default value is a column/vector named \code{CNSR}}
}
\value{
Object of class 'Surv'
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Create a survival object (e.g. \code{survival::Surv()}), which is usually used as
the response variable in survival models.
The function is designed to leverage the
\href{https://www.cdisc.org/standards/foundational/adam/adam-basic-data-structure-bds-time-event-tte-analyses-v1-0}{CDISC ADaM ADTTE data model}
and more explicitly the conventions and controlled vocabulary of the data model.

The \code{AVAL} and \code{CNSR} arguments are passed to
\code{survival::Surv(time = AVAL, event = 1 - CNSR, type = 'right', origin = 0)}.
}
\examples{
survival::survfit(visR::Surv_CDISC() ~ SEX, data = adtte)
adtte \%>\%
 visR:: estimate_KM(formula = visR::Surv_CDISC() ~ SEX)

# When using CDSIC data, you can specify arguments or rely on defaults
with(adtte, visR::Surv_CDISC(AVAL, CNSR)) \%>\% head()
with(adtte, visR::Surv_CDISC()) \%>\% head()
}
\seealso{
\code{\link[survival:Surv]{survival::Surv()}}, \code{\link[=estimate_KM]{estimate_KM()}}
}
