% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_download}
\alias{table_download}
\title{Create a caption for any table}
\usage{
table_download(df, filename = NULL, format = "csv", button = "Download data")
}
\arguments{
\item{df}{The df}

\item{filename}{The filename}

\item{format}{What format. One of "txt", "tsv", "csv", "xlsx"}

\item{button}{What does the download button say}
}
\description{
Create a caption for any table
}
