% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{tableone}
\alias{tableone}
\title{Summary Table (Table One)}
\usage{
tableone(
  data,
  title,
  datasource,
  strata = NULL,
  summary_function = summarize_tab1,
  ...
)
}
\arguments{
\item{data}{The dataframe or tibble to visualise}

\item{title}{Table title to include in the rendered table}

\item{datasource}{String specifying the datasource underlying the data set}

\item{strata}{Character vector with column names to use for
stratification in the summary table. Default: NULL}

\item{summary_function}{A function to summarize variables of different types.
Pre-implemented functions are summarize and summarize_tab1}

\item{...}{Pass options to render_table}
}
\description{
Create and render a summary table for a dataset

Calculate summary statistics and present them in a formatted table
}
