% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pipe.R
\name{the_lhs}
\alias{the_lhs}
\title{Find the "lhs" in the pipeline}
\usage{
the_lhs()
}
\value{
Left-hand sided symbol as string in the magrittr pipe.
}
\description{
This function finds the left-hand sided symbol in a magrittr pipe and returns it as a character.
}
\examples{
blah <- function(x) the_lhs()
adtte \%>\%
  blah()
}
\references{
\url{https://github.com/tidyverse/magrittr/issues/115#issuecomment-173894787}
}
