% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_create_tableone.R
\name{vr_create_tableone}
\alias{vr_create_tableone}
\title{Create Summary Table (also known as Table 1)}
\usage{
vr_create_tableone(
  dat,
  group_columns = NULL,
  overall = TRUE,
  summary_function = vr_summarize_tab1
)
}
\arguments{
\item{dat}{The data set to summarize as data frame or tibble}

\item{group_columns}{Stratifying/Grouping variable name(s) as character vector.
If NULL, only overall results are returned}

\item{overall}{If TRUE, the summary statistics for the overall data set are also calculated}

\item{summary_function}{A function defining summary statistics for numeric and categorical values}
}
\description{
Create a summary table of descriptive statistics from a data frame or tibble.

By default the following summary stats are calculated:
\itemize{
\item Numeric variables: mean, min, 25th-percentile, median, 75th-percentile, maximum, standard deviation
\item Factor variables: proportion of each factor level in the overall data set
\item Default: number of unique values and number of missing values
}
}
\details{
It is possible to provide your own summary function. Please have a loot at vr_summary for inspiration.
}
\note{
All columns in the table will be summarized.
If only some columns shall be used, please select only those
variables prior to creating the summary table by using dplyr::select()
}
\examples{
library(survival)
library(dplyr)
ovarian \%>\% 
select(-fustat) \%>\% 
  mutate(age_group = factor(case_when(age <= 50 ~ "<= 50 years",
                                      age <= 60 ~ "<= 60 years",
                                      age <= 70 ~ "<= 70 years",
                                      TRUE ~ "> 70 years")),
         rx = factor(rx),
         ecog.ps = factor(ecog.ps)) \%>\% 
  select(age, age_group, everything()) \%>\% 
  vr_create_tableone()
}
