% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_kaplan_meier_summary.R
\name{vr_kaplan_meier_summary}
\alias{vr_kaplan_meier_summary}
\title{Plot Kaplan-Meier Summary Table for Existing}
\usage{
vr_kaplan_meier_summary(data, equation)
}
\arguments{
\item{data}{Dataset as dataframe/tibble containing with event data (one row per subject)}

\item{equation}{Formula to create survival object}
}
\value{
List containing two summary tables: 1) Summary Table with persons at risk, events, median survival times along with 95\% CIs over strata and 2) Summary table with test of equality over strata
}
\description{
Plot Kaplan-Meier Summary Table for Existing
}
