% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_kaplan_meier_summary.R
\name{vr_kaplan_meier_summary}
\alias{vr_kaplan_meier_summary}
\title{Plot Kaplan-Meier Summary Table for Existing}
\usage{
vr_kaplan_meier_summary(data, equation)
}
\arguments{
\item{data}{Dataset as dataframe/tibble containing with event data (one row per subject)}

\item{equation}{Formula to create survival object}
}
\value{
List containing two summary tables: 1) Summary Table with persons at risk, events, median survival times along with 95\% CIs over strata and 2) Summary table with test of equality over strata
}
\description{
TODO: Check the warnings thrown in the examples.
}
\examples{
#' # Create interim data models
library(survival)
library(survminer)
library(ggplot2)
library(dplyr)
data("veteran")
data <-  veteran \%>\% 
    mutate(trt = as.factor(case_when(
       trt == 1 ~ "standard therapy", 
       trt == 2 ~ "test chemotherapy"
   )))

equation <- "survival::Surv(time, status) ~ trt"
vr_kaplan_meier_summary(data, equation) 

# Example
output <- vr_kaplan_meier_summary(data=BRCAOV.survInfo, equation="Surv(times, patient.vital_status) ~ admin.disease_code")
output[1] # Summary Table with persons at risk, events, median survival times along with 95\% CIs over strata
output[2] # Summary table with test of equality over strata

}
