% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plt_kaplan_meier.R
\name{vr_plt_kaplan_meier}
\alias{vr_plt_kaplan_meier}
\title{Plot Kaplan-Meier Curve for Existing Tidy Survival Object}
\usage{
vr_plt_kaplan_meier(
  broom_object,
  title = "",
  abbreviations = NULL,
  variable_definitions = NULL,
  N = NULL,
  N_unit = "patients",
  time_unit = "days",
  data_source = NULL,
  estimate_name = "survival probability"
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Plot Kaplan-Meier Curve for Existing Tidy Survival Object
}
