% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plt_kaplan_meier.R
\name{vr_plt_kaplan_meier}
\alias{vr_plt_kaplan_meier}
\title{Plot Kaplan-Meier Curve for Existing Tidy Survival Object}
\usage{
vr_plt_kaplan_meier(
  broom_object,
  title = "",
  abbreviations = NULL,
  variable_definitions = NULL,
  N = NULL,
  N_unit = "patients",
  time_unit = "days",
  data_source = NULL,
  estimate_name = "survival probability"
)
}
\arguments{
\item{broom_object, }{}

\item{title}{= "",}

\item{abbreviations}{= NULL,}

\item{variable_definitions}{= NULL,}

\item{N}{= NULL,}

\item{N_unit}{= "patients",}

\item{time_unit}{= "days",}

\item{data_source}{= NULL,}

\item{estimate_name}{= "survival probability"}
}
\value{
ggplot object
}
\description{
TODO: Define and describe the purpose of this function.
}
\examples{
# TODO: Define an example for this function
# Create interim data models
library(survival)
library(ggplot2)
library(dplyr)
data("veteran")
data <-  veteran \%>\% 
    mutate(trt = as.factor(case_when(
       trt == 1 ~ "standard therapy", 
       trt == 2 ~ "test chemotherapy"
   )))

equation <- "survival::Surv(time, status) ~ trt"
risk_table_data <- vr_est_km_risk_table(data, equation)
time_unit = "days"
broom_object <- vr_est_kaplan_meier(data, equation)
risk_table_data <- vr_est_km_risk_table(data, equation)

# Create plot
table <- vr_plt_km_risk_table(risk_table_data,
                              time_unit = time_unit)
                              
##not run TODO: Check why this example is failing
#vr_plt_kaplan_meier(
# broom_object, 
#  N = "Patients", 
#  time_unit = "days", 
#  #'  data_source = "this is the data source label"
# )
}
