% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_plt_km_risk_table.R
\name{vr_plt_km_risk_table}
\alias{vr_plt_km_risk_table}
\title{Plot (Kaplan-Meier) Risk Table for an existing tidy risk object}
\usage{
vr_plt_km_risk_table(risk_table_data, time_unit)
}
\arguments{
\item{time_unit}{}

\item{data}{}
}
\value{
ggplot object for the risk table
}
\description{
Plot (Kaplan-Meier) Risk Table for an existing tidy risk object
}
\examples{
# Create interim data models
library(survival)
library(ggplot2)
library(dplyr)
data("veteran")
data <-  veteran \%>\% 
    mutate(trt = as.factor(case_when(
       trt == 1 ~ "standard therapy", 
       trt == 2 ~ "test chemotherapy"
   )))

equation <- "survival::Surv(time, status) ~ trt"
risk_table_data <- vr_est_km_risk_table(data, equation)
time_unit = "days"

# Create plot
vr_plt_km_risk_table(risk_table_data,
                     time_unit = time_unit)

}
