% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize_tab1.numeric}
\alias{vr_summarize_tab1.numeric}
\title{Create variable summary for numeric variables}
\usage{
\method{vr_summarize_tab1}{numeric}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Create variable summary for numeric variables
}
