% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_utils.R
\name{vr_summarize_tab1}
\alias{vr_summarize_tab1}
\title{Create abbreviated variable summary for table1}
\usage{
vr_summarize_tab1(x)
}
\arguments{
\item{x}{a vector to be summarized}
}
\description{
This function creates summaries combines multiple summary measures in a single formatted string.
}
