% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_table_one.R
\name{vr_table_one}
\alias{vr_table_one}
\title{Summary Table (Table One)}
\usage{
vr_table_one(
  dat,
  title,
  data_source,
  caption = "",
  group_columns = NULL,
  summary_function = vr_summarize_tab1,
  ...
)
}
\arguments{
\item{dat}{The data set to summarize as data frame or tibble}

\item{title}{Table title to include in the rendered table}

\item{data_source}{String specifying the datas ource underlying the data set}

\item{caption}{Table caption to include in the rendered table}

\item{group_columns}{Stratifying/Grouping variable name(s) as character vector.
If NULL, only overall results are returned}

\item{summary_function}{A function to summarize variables of different types.
Pre-implemented functions are vr_summarize and vr_summarize_tab1}

\item{...}{Pass options to vr_render_tableone}
}
\description{
Create and render a summary table for a data set

Calculate summary statistics and present them in a formatted table
}
