% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_table_one.R
\name{vr_table_one}
\alias{vr_table_one}
\title{Summary Table (Table One)}
\usage{
vr_table_one(
  data,
  title,
  datasource,
  caption = "",
  groupCols = NULL,
  summary_function = vr_summarize_tab1,
  ...
)
}
\arguments{
\item{data}{The dataframe or tibble to visualise}

\item{title}{Table title to include in the rendered table}

\item{datasource}{String specifying the datasource underlying the data set}

\item{caption}{Table caption to include in the rendered table}

\item{groupCols}{Character vector with column names to use for stratification in the summary table. Default: NULL}

\item{summary_function}{A function to summarize variables of different types. Pre-implemented functions are vr_summarize and vr_summarize_tab1}

\item{engine}{Rendering engine to use for creating output. Currently implemented are gt, kable and jquery datatable (DT)}
}
\description{
Create and render a summary table for a dataset

Calculate summary statistics and present them in a formatted table
}
