% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_tidy_rbest.R
\name{vr_tidy_rbest}
\alias{vr_tidy_rbest}
\title{Returns a tidied RBesT gMAP object}
\usage{
vr_tidy_rbest(x, prob = 0.95)
}
\arguments{
\item{x}{RBesT gMAP object}

\item{prob}{probability range for uncertainty interval}
}
\value{
td tidied tibble
}
\description{
This is an experimental function that may be developed over time.

\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
\donttest{
# Commenting out to not depend on RBest - issue with cran availability
# library(RBesT)
# example(crohn)
# vr_tidy_rbest(map_crohn)

# map_crohn \%>\%
# vr_tidy_rbest()
}

}
